RHSTOR-7671:TC to verify ODF upgrade prevent status when Ceph Health WARN
Test that ODF upgrade is blocked when Ceph cluster health is in WARN state.
This Test Verifies:
1. OCS OperatorCondition CR: Upgradeable=False
2. ODF OperatorCondition CR: Upgradeable=False
3. Alert ODFOperatorNotUpgradeable
